requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiochclient[aiohttp-speedups]==2.1.0
    # via -r requirements.in
aiocontextvars==0.2.2
    # via -r requirements.in
aiodns==2.0.0
    # via aiochclient
aiohttp==3.7.4
    # via
    #   aiochclient
    #   discord-py
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
beautifulsoup4==4.10.0
    # via -r requirements.in
cachetools==4.1.1
    # via google-auth
cchardet==2.1.7
    # via aiochclient
certifi==2020.11.8
    # via
    #   requests
    #   sentry-sdk
cffi==1.14.3
    # via
    #   pycares
    #   pynacl
chardet==3.0.4
    # via
    #   aiohttp
    #   requests
ciso8601==2.1.3
    # via aiochclient
cycler==0.10.0
    # via matplotlib
defusedxml==0.7.1
    # via -r requirements.in
deprecated==1.2.13
    # via redis
discord-py[voice]==2.0.1
    # via -r requirements.in
fonttools==4.32.0
    # via matplotlib
google-api-core==1.23.0
    # via google-api-python-client
google-api-python-client==2.33.0
    # via -r requirements.in
google-auth==1.23.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.1.0
    # via google-api-python-client
googleapis-common-protos==1.52.0
    # via google-api-core
httplib2==0.19.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==2.10
    # via
    #   requests
    #   yarl
imagehash==4.2.1
    # via -r requirements.in
kiwisolver==1.3.1
    # via matplotlib
matplotlib==3.5.1
    # via -r requirements.in
multidict==4.7.6
    # via
    #   aiohttp
    #   yarl
numpy==1.21.5
    # via
    #   -r requirements.in
    #   imagehash
    #   matplotlib
    #   pywavelets
packaging==21.3
    # via
    #   matplotlib
    #   redis
pillow==8.4.0
    # via
    #   -r requirements.in
    #   imagehash
    #   matplotlib
protobuf==3.14.0
    # via
    #   google-api-core
    #   googleapis-common-protos
psutil==5.9.0
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
ptvsd==4.3.2
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycares==3.1.1
    # via aiodns
pycparser==2.20
    # via cffi
pynacl==1.3.0
    # via discord-py
pyparsing==2.4.7
    # via
    #   httplib2
    #   matplotlib
    #   packaging
python-dateutil==2.8.1
    # via matplotlib
pytube==12.0.0
    # via -r requirements.in
pytz==2020.4
    # via google-api-core
pywavelets==1.1.1
    # via imagehash
redis==4.1.0
    # via -r requirements.in
requests==2.25.0
    # via google-api-core
rsa==4.7
    # via google-auth
scipy==1.5.4
    # via imagehash
sentry-sdk==1.5.10
    # via -r requirements.in
six==1.15.0
    # via
    #   cycler
    #   google-api-core
    #   google-auth
    #   google-auth-httplib2
    #   imagehash
    #   protobuf
    #   pynacl
    #   python-dateutil
soupsieve==2.0.1
    # via beautifulsoup4
sqlalchemy==1.4.29
    # via -r requirements.in
sqlparse==0.4.2
    # via aiochclient
typing-extensions==3.10.0.0
    # via aiohttp
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.5
    # via
    #   requests
    #   sentry-sdk
wrapt==1.14.0
    # via deprecated
yarl==1.5.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
